+++ CHANGELOG.md
@@ -103,3 +103,7 @@
   - add `Akeneo\Tool\Component\StorageUtils\Saver\SaverInterface` (twice) and `Akeneo\Tool\Component\StorageUtils\Saver\BulkSaverInterface`
 - Change interface `Akeneo\Tool\Bundle\RuleEngineBundle\Model\RuleDefinitionInterface` to add `isEnabled` and `setEnabled` methods
 - Change interface `Akeneo\Tool\Bundle\RuleEngineBundle\Repository\RuleDefinitionRepositoryInterface` to rename `findAllOrderedByPriority` method by `findEnabledOrderedByPriority`
+- Change constructor of `Akeneo\Pim\Automation\RuleEngine\Bundle\Controller\InternalApi\GetCategoriesController` to replace `ObjectFilterInterface $objectFilter` by `CollectionFilterInterface $collectionFilter`
+- Change constructor of `Akeneo\Test\Pim\Automation\RuleEngine\Integration\Context\AssociationContext` to replace `IdentifiableObjectRepositoryInterface $productRepository` by `ProductRepositoryInterface $productRepository`
+- Remove `Akeneo\Pim\Enrichment\AssetManager\Bundle\Analytics\CountAssetFamilies` class
+- Remove `Akeneo\Pim\Enrichment\ReferenceEntity\Bundle\Analytics\CountReferenceEntities` class